Module type Tjr_mmap__Mmap_intf.S

To write to/from the mmap, use the sub function to get the subarray of interest. NOTE that because the map may grow when asking for a sub which extends beyond the current length, a remap may occur. Thus, subs should not be reused or stored, since they may refer to an out of date slice for an old mapping. For this reason, subs are not thread safe (a new thread may unmap a sub being used by an older thread).

type ('a, 'b) t

val bytes_per_elt : ('a, 'b) Tjr_mmap.Util.Type_abbrevs.kind -> int

Number of bytes used to store each elt

val of_fd : Unix.file_descr -> ('a, 'b) Tjr_mmap.Util.Type_abbrevs.kind -> ('a, 'b) t

val length : (_, _) t -> int

Length of the mmap in terms of number of elts; should match the fstat size (taking bytes_per_elt into account)

val sub : ('a, 'b) t -> off:int -> len:int -> ('a, 'b, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

Subarray from offset of the given length

val msync : (_, _) t -> unit

Perform an msync to flush data to disk

val close : (_, _) t -> unit